+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 18.06.7
 - tags in the `Move tag to...` context menu of the tag list are now sorted by name
  (for [#947](https://github.com/pbek/QOwnNotes/issues/947))
+- if a code block is inserted in an empty line then automatically a multi-line
+  code block will be inserted instead of an inline code block
+  (for [#912](https://github.com/pbek/QOwnNotes/issues/912))
 
 ## 18.06.6
 - improved setting the size of the settings dialog if it was opened for the